wireless-regdb: Correct antenna gain
The original db.txt came from a program which scaped the ath9k
regulatory framework, in it a value of 6 was used for antenna
gain. The value for antenna gain in ath9k is twice the actual
value, so a value of 6 is actually equivalent to 3 dBi. Lets
correct this in db.txt

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>