wireless-regdb: KR few reg updates
Korea (KR) allows the usage a channel with center of
freq on 5260 MHz. It also specifically disables these
channels though:

128 - 5640 MHz
132 - 5660 MHz
136 - 5680 MHz
140 - 5700 MHz

We refered to the Korean rules regulations, the doc is in
Korean and I am not sure if I can post it but just in case
Atheros refers to this doc internally as "MIC 2007-80 20071017".

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>