wireless-regdb: update US rules for 5600 MHz - 5650 MHz
The FCC is trying to assist airports that use Terminal Doppler
Weather Radar (TDWR) systems in avoiding interference with
some outdoor wireless systems operating in the 5.4 GHz
(5470 MHz - 5725 MHz) band. One of the things they have decided
on is to disallow operation on the 5600 MHz - 5650 MHz frequency
range inclusive. What this means in practice is 5 GHz 802.11
devices programmed to operate in the US will have these
channels now disabled:

 * 5600 MHz [120] (disabled)
 * 5620 MHz [124] (disabled)
 * 5640 MHz [128] (disabled)

Channel on 5660 MHz (132) will be left enabled as the channel
bandwidth is 20 MHz for legacy 802.11 networks but note that
HT40- will be disallowed for that channel. The HT40+ and HT40-
allowed channel map obviously changes. This is computed
dynamically on cfg80211 based on your allowed channel list.
For a US based device this is what you will get now
(excerpt from on /sys/kernel/debug/ieee80211/phy0):

 .. etc ..
5540 HT40 -+
5560 HT40 -+
5580 HT40 -
5600 Disabled
5620 Disabled
5640 Disabled
5660 HT40  +
5680 HT40 -+
 .. etc ..

The details on which this patch is based on is documented on
the FCC Knowledge based Publication Number: 443999 [1]

[1] https://fjallfoss.fcc.gov/oetcf/kdb/forms/FTSSearchResultPage.cfm?switch=P&id=41732

Cc: Michael Green <Michael.Green@atheros.com>
Cc: Senthil Balasubramanian <senthilkumar@atheros.com>
Cc: David Quan <David.Quan@atheros.com>
Cc: Vivek Natarajan <Vivek.Natarajan@atheros.com>
Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>