wireless-regdb: add Aruba (AW)
We were unable to find published rules for Aruba
for regulatory purposes but given that it falls under
the same european region we have mapped Aruba to the same
region as other european countries and give it the same
rules. A device was tested for ETSI certification with
these rules and passed so I think this is a best due dillegence
as we can do to test regulatory rules for Aruba.

Reported-by: Jeffrey Baker <jwbaker@gmail.com>
Cc: Michael Green <Michael.Green@atheros.com>
Cc: Vivek Natarajan <Vivek.Natarajan@atheros.com>
Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>