dbparse: remove usable bandwidth check
The kernel uses the bandwidth this way: any frequency
rule that is required for any of the 20 MHz subchannels
of a given channel must allow the total bandwidth.
Therefore, this check is wrong -- a frequency rule may
need to be specified with higher bandwidth than it has
to allow to form e.g. HT40 out of subchannels.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>