regdb: remove wrong 10 MHz rule for JP
The kernel currently doesn't actually check the
maximum bandwidth for 20 MHz channels (which is
arguably a bug), but if it were to check then
this rule would prevent channel 184 from being
used which is wrong.

The rule does, however, prevent HT40 from being
used when the primary or secondary channel is
channel 184 as the kernel checks there. This is
wrong though, it should be allowed.

Remove the rule.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>